+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project are documented in this file.
 
 ## Unreleased
 
+### Slint Language
+
+ - Added `clamp` function that takes a `value`, `minimum` and `maxium` and will return
+   `maximum` if `value > maximum`, `minimum` if `value < minimum` or `value` otherwise.
+
 ## [1.3.0] - 2023-11-10
 
 ### General